Revert change that did markdown conversion last to avoid escaped markdown. Add regression test to make guarantee about markdown conversion.
